backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
openai
langchain
langchain-community
langchain-openai
faiss-cpu
spacy
keybert
googletrans==4.0.0-rc1
langdetect
pymupdf
pandas
openpyxl
python-docx
scikit-learn
tiktoken
numpy
